+++ lib/WorseReflection/Bridge/TolerantParser/Reflection/AbstractReflectionClass.php
@@ -69,6 +69,7 @@ abstract class AbstractReflectionClass extends AbstractReflectedNode implements 
                 }
 
                 $traitAlias = $traitImport->getAlias($method->name());
+                dump('HERE', $this->name(), $trait->name(), $method->name());
 
                 $virtualMethod = VirtualReflectionMethod::fromReflectionMethod($trait->methods()->get($traitAlias->originalName()))
                     ->withName($traitAlias->newName())
