+++ lib/WorseReflection/Bridge/TolerantParser/Reflection/ReflectionDeclaredConstant.php
@@ -7,7 +7,7 @@ use Microsoft\PhpParser\Node\Expression\ArgumentExpression;
 use Microsoft\PhpParser\Node\Expression\CallExpression;
 use Microsoft\PhpParser\Node\StringLiteral;
 use Phpactor\WorseReflection\Core\DocBlock\DocBlock;
-use Phpactor\WorseReflection\Core\Inference\Frame;
+use Phpactor\WorseReflection\Core\Inference\ConcreteFrame;
 use Phpactor\WorseReflection\Core\Name;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionDeclaredConstant as PhpactorReflectionDeclaredConstant;
 use Phpactor\TextDocument\TextDocument;
@@ -35,7 +35,7 @@ class ReflectionDeclaredConstant extends AbstractReflectedNode implements Phpact
 
     public function type(): Type
     {
-        return $this->serviceLocator->nodeContextResolver()->resolveNode(new Frame(), $this->value)->type();
+        return $this->serviceLocator->nodeContextResolver()->resolveNode(new ConcreteFrame(), $this->value)->type();
     }
 
     public function sourceCode(): TextDocument
