+++ lib/WorseReflection/Bridge/TolerantParser/Reflection/ReflectionEnumCase.php
@@ -11,7 +11,7 @@ use Phpactor\WorseReflection\Core\Reflection\ReflectionClassLike;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionMember;
 use Phpactor\WorseReflection\Core\Type;
 use Phpactor\WorseReflection\Core\ServiceLocator;
-use Phpactor\WorseReflection\Core\Inference\Frame;
+use Phpactor\WorseReflection\Core\Inference\ConcreteFrame;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionEnumCase as CoreReflectionEnumCase;
 use Phpactor\WorseReflection\Core\TypeFactory;
 use Phpactor\WorseReflection\Core\Type\MissingType;
@@ -86,7 +86,7 @@ class ReflectionEnumCase extends AbstractReflectionClassMember implements CoreRe
         return $this->serviceLocator()
                     ->nodeContextResolver()
                     ->resolveNode(
-                        new Frame(),
+                        new ConcreteFrame(),
                         $this->node->assignment
                     )->type();
     }
