+++ lib/WorseReflection/Core/Virtual/VirtualReflectionFunction.php
@@ -6,6 +6,7 @@ use Phpactor\TextDocument\ByteOffsetRange;
 use Phpactor\TextDocument\TextDocumentBuilder;
 use Phpactor\WorseReflection\Core\DocBlock\DocBlock;
 use Phpactor\WorseReflection\Core\DocBlock\PlainDocblock;
+use Phpactor\WorseReflection\Core\Inference\ConcreteFrame;
 use Phpactor\WorseReflection\Core\Inference\Frame;
 use Phpactor\WorseReflection\Core\Name;
 use Phpactor\WorseReflection\Core\NodeText;
@@ -39,7 +40,7 @@ final class VirtualReflectionFunction implements ReflectionFunction
         return new self(
             $range,
             NodeText::fromString(''),
-            new Frame(),
+            new ConcreteFrame(),
             new PlainDocblock(),
             new DummyReflectionScope(),
             TypeFactory::undefined(),
