+++ lib/WorseReflection/Tests/Unit/Core/Inference/FrameTest.php
@@ -4,6 +4,7 @@ namespace Phpactor\WorseReflection\Tests\Unit\Core\Inference;
 
 use Phpactor\WorseReflection\Core\Inference\Assignments;
 use PHPUnit\Framework\TestCase;
+use Phpactor\WorseReflection\Core\Inference\ConcreteFrame;
 use Phpactor\WorseReflection\Core\Inference\Frame;
 use Phpactor\WorseReflection\Core\Inference\LocalAssignments;
 use Phpactor\WorseReflection\Core\Inference\PropertyAssignments;
@@ -15,7 +16,7 @@ class FrameTest extends TestCase
      */
     public function testAssignments(): void
     {
-        $frame = new Frame();
+        $frame = new ConcreteFrame();
         $this->assertInstanceOf(LocalAssignments::class, $frame->locals());
         $this->assertInstanceOf(PropertyAssignments::class, $frame->properties());
     }
