+++ lib/WorseReflection/Tests/Unit/Core/Inference/NodeReflectorTest.php
@@ -4,6 +4,7 @@ namespace Phpactor\WorseReflection\Tests\Unit\Core\Inference;
 
 use PHPUnit\Framework\TestCase;
 use Phpactor\WorseReflection\Core\Exception\CouldNotResolveNode;
+use Phpactor\WorseReflection\Core\Inference\ConcreteFrame;
 use Phpactor\WorseReflection\Core\ServiceLocator;
 use Phpactor\WorseReflection\Core\Inference\NodeReflector;
 use Phpactor\WorseReflection\Core\Inference\Frame;
@@ -18,7 +19,7 @@ class NodeReflectorTest extends TestCase
     {
         $this->expectException(CouldNotResolveNode::class);
         $this->expectExceptionMessage('Did not know how');
-        $frame = new Frame();
+        $frame = new ConcreteFrame();
         $locator = $this->prophesize(ServiceLocator::class);
         $nodeReflector = new NodeReflector($locator->reveal());
